+++ eras/dijkstra/impl/src/Cardano/Ledger/Dijkstra/Rules/SubEntities.hs
@@ -103,6 +103,9 @@ instance InjectRuleFailure "SUBENTITIES" SubEntitiesPredFailure DijkstraEra
 instance InjectRuleFailure "SUBENTITIES" DijkstraSubCertsPredFailure DijkstraEra where
   injectFailure = SubCertsFailure
 
+instance InjectRuleFailure "SUBENTITIES" Conway.ConwayCertsPredFailure DijkstraEra where
+  injectFailure = SubCertsFailure . injectFailure @"SUBCERTS"
+
 instance InjectRuleFailure "SUBENTITIES" Conway.ConwayLedgerPredFailure DijkstraEra where
   injectFailure = conwayToDijkstraSubEntitiesPredFailure