+++ eras/dijkstra/impl/testlib/Test/Cardano/Ledger/Dijkstra/TreeDiff.hs
@@ -263,18 +263,26 @@ instance
   ToExpr (Event (EraRule "SUBCERT" era)) =>
   ToExpr (DijkstraSubCertsEvent era)
 
+instance
+  ToExpr (PredicateFailure (EraRule "SUBCERTS" era)) =>
+  ToExpr (SubEntitiesPredFailure era)
+
+instance
+  ToExpr (Event (EraRule "SUBCERTS" era)) =>
+  ToExpr (SubEntitiesEvent era)
+
 instance ToExpr (DijkstraSubDelegPredFailure era)
 
 instance
   ( ToExpr (PredicateFailure (EraRule "SUBGOV" era))
-  , ToExpr (PredicateFailure (EraRule "SUBCERTS" era))
+  , ToExpr (PredicateFailure (EraRule "SUBENTITIES" era))
   , ToExpr (PredicateFailure (EraRule "SUBUTXOW" era))
   ) =>
   ToExpr (DijkstraSubLedgerPredFailure era)
 
 instance
   ( ToExpr (Event (EraRule "SUBGOV" era))
-  , ToExpr (Event (EraRule "SUBCERTS" era))
+  , ToExpr (Event (EraRule "SUBENTITIES" era))
   , ToExpr (Event (EraRule "SUBUTXOW" era))
   ) =>
   ToExpr (DijkstraSubLedgerEvent era)